+++ library/cpp/yt/error/error.h
@@ -1,8 +1,6 @@
 #pragma once
 
-#include "public.h"
-
-#include <yt/yt/core/ytree/public.h>
+#include <library/cpp/yt/error/error_code.h>
 
 #include <library/cpp/yt/threading/public.h>
 
@@ -440,5 +438,5 @@ auto RunNoExcept(F&& functor, As&&... args) noexcept -> decltype(functor(std::fo
 } // namespace NYT
 
 #define STRIPPED_ERROR_INL_H_
-#include "stripped_error-inl.h"
+#include "error-inl.h"
 #undef STRIPPED_ERROR_INL_H_
