+++ library/cpp/yt/error/error_attributes.cpp
@@ -1,30 +1,70 @@
 #include "error_attributes.h"
 
+#include "error.h"
+#include "error_code.h"
+
 namespace NYT {
 
 ////////////////////////////////////////////////////////////////////////////////
 
-TErrorAttributes::TErrorAttributes(void* attributes)
-    : Attributes_(attributes)
-{ }
+std::vector<TErrorAttributes::TKey> TErrorAttributes::ListKeys() const
+{
+    std::vector<TString> keys;
+    keys.reserve(Map_.size());
+    for (const auto& [key, value] : Map_) {
+        keys.push_back(key);
+    }
+    return keys;
+}
+
+std::vector<TErrorAttributes::TKeyValuePair> TErrorAttributes::ListPairs() const
+{
+    std::vector<TKeyValuePair> pairs;
+    pairs.reserve(Map_.size());
+    for (const auto& pair : Map_) {
+        pairs.push_back(pair);
+    }
+    return pairs;
+}
+
+TErrorAttributes::TValue TErrorAttributes::FindValue(TStringBuf key) const
+{
+    auto it = Map_.find(key);
+    return it == Map_.end() ? TValue{} : it->second;
+}
+
+void TErrorAttributes::SetValue(const TKey& key, const TValue& value)
+{
+    Map_[key] = value;
+}
+
+void TErrorAttributes::SetAttribute(const TErrorAttribute& attribute)
+{
+    SetValue(attribute.Key, attribute.Value);
+}
+
+bool TErrorAttributes::Remove(const TKey& key)
+{
+    return Map_.erase(key) > 0;
+}
+
+TErrorAttributes::TValue TErrorAttributes::GetValue(TStringBuf key) const
+{
+    auto result = FindValue(key);
+    if (!result) {
+        ThrowNoSuchAttributeException(key);
+    }
+    return result;
+}
 
 void TErrorAttributes::Clear()
 {
-    for (const auto& key : ListKeys()) {
-        Remove(key);
-    }
-}
-
-TErrorAttributes::TValue TErrorAttributes::GetYsonAndRemove(const TKey& key)
-{
-    auto result = GetYson(key);
-    Remove(key);
-    return result;
+    Map_.clear();
 }
 
 bool TErrorAttributes::Contains(TStringBuf key) const
 {
-    return FindYson(key).operator bool();
+    return Map_.contains(key);
 }
 
 bool operator == (const TErrorAttributes& lhs, const TErrorAttributes& rhs)
@@ -55,6 +95,61 @@ bool operator == (const TErrorAttributes& lhs, const TErrorAttributes& rhs)
     }
 
     return true;
+}
+
+////////////////////////////////////////////////////////////////////////////////
+
+namespace {
+
+bool IsSpecialCharacter(char ch)
+{
+    return ch == '\\' || ch == '/' || ch == '@' || ch == '*' || ch == '&' || ch == '[' || ch == '{';
+}
+
+// AppendYPathLiteral.
+void DoFormatAttributeKey(TStringBuilderBase* builder, TStringBuf value)
+{
+    constexpr char asciiBegin = 32;
+    constexpr char asciiEnd = 127;
+    builder->Preallocate(value.length() + 16);
+    for (unsigned char ch : value) {
+        if (IsSpecialCharacter(ch)) {
+            builder->AppendChar('\\');
+            builder->AppendChar(ch);
+        } else if (ch < asciiBegin || ch > asciiEnd) {
+            builder->AppendString(TStringBuf("\\x"));
+            builder->AppendChar(IntToHexLowercase[ch >> 4]);
+            builder->AppendChar(IntToHexLowercase[ch & 0xf]);
+        } else {
+            builder->AppendChar(ch);
+        }
+    }
+}
+
+} // namespace
+
+////////////////////////////////////////////////////////////////////////////////
+
+[[noreturn]] void TErrorAttributes::ThrowCannotParseAttributeException(TStringBuf key, const std::exception& ex)
+{
+    THROW_ERROR_EXCEPTION(
+        "Error parsing attribute %Qv",
+        key)
+        << ex;
+}
+
+[[noreturn]] void TErrorAttributes::ThrowNoSuchAttributeException(TStringBuf key)
+{
+    auto formatAttributeKey = [] (auto key) {
+        TStringBuilder builder;
+        DoFormatAttributeKey(&builder, key);
+        return builder.Flush();
+    };
+
+    THROW_ERROR_EXCEPTION(
+        /*NYTree::EErrorCode::ResolveError*/ TErrorCode{500},
+        "Attribute %Qv is not found",
+        formatAttributeKey(key));
 }
 
 ////////////////////////////////////////////////////////////////////////////////
