+++ library/cpp/yt/error/origin_attributes.cpp
@@ -1,4 +1,5 @@
 #include "origin_attributes.h"
+#include "error_attributes.h"
 
 #include <library/cpp/yt/assert/assert.h>
 
@@ -103,6 +104,48 @@ TString FormatOrigin(const TOriginAttributes& attributes)
             }
             FormatValue(builder, threadName, "v");
         }));
+}
+
+////////////////////////////////////////////////////////////////////////////////
+
+TOriginAttributes ExtractFromDictionary(TErrorAttributes* attributes)
+{
+    using TFunctor = TOriginAttributes(*)(TErrorAttributes*);
+
+    if (auto strong = NGlobal::GetErasedVariable(ExtractFromDictionaryTag)) {
+        return strong->AsConcrete<TFunctor>()(attributes);
+    }
+
+    return ExtractFromDictionaryDefault(attributes);
+}
+
+////////////////////////////////////////////////////////////////////////////////
+
+TOriginAttributes ExtractFromDictionaryDefault(TErrorAttributes* attributes)
+{
+    TOriginAttributes result;
+    if (attributes == nullptr) {
+        return result;
+    }
+
+    // TODO(arkady-e1ppa): Try using std::string here.
+    static const TString HostKey("host");
+    result.HostHolder = TSharedRef::FromString(attributes->GetAndRemove(HostKey, TString()));
+    result.Host = result.HostHolder.empty() ? TStringBuf() : TStringBuf(result.HostHolder.Begin(), result.HostHolder.End());
+
+    static const TString DatetimeKey("datetime");
+    result.Datetime = attributes->GetAndRemove(DatetimeKey, TInstant());
+
+    static const TString PidKey("pid");
+    result.Pid = attributes->GetAndRemove(PidKey, TProcessId{});
+
+    static const TString TidKey("tid");
+    result.Tid = attributes->GetAndRemove(TidKey, NThreading::InvalidThreadId);
+
+    static const TString ThreadNameKey("thread");
+    result.ThreadName = attributes->GetAndRemove<TString>(ThreadNameKey, TString());
+
+    return result;
 }
 
 } // namespace NDetail
