+++ library/cpp/yt/error/ya.make
@@ -10,11 +10,20 @@ PEERDIR(
     library/cpp/yt/threading
     library/cpp/yt/string
     library/cpp/yt/yson_string # TODO(arkady-e1ppa): eliminate
+    library/cpp/yt/logging # TODO(arkady-e1ppa): Consider logging error_code crashes to stderr and drop this dep.
+
+    util
 )
 
 SRCS(
+    error.cpp
     error_attributes.cpp
+    error_code.cpp
     origin_attributes.cpp
 )
 
 END()
+
+RECURSE_FOR_TESTS(
+    unittests
+)
