+++ library/cpp/yt/string/string.cpp
@@ -359,4 +359,19 @@ TStringBuf FormatBool(bool value)
 
 ////////////////////////////////////////////////////////////////////////////////
 
+void TruncateStringInplace(TString* string, int lengthLimit, TStringBuf truncatedSuffix)
+{
+    if (std::ssize(*string) > lengthLimit) {
+        *string = Format("%v%v", string->substr(0, lengthLimit), truncatedSuffix);
+    }
+}
+
+TString TruncateString(TString string, int lengthLimit, TStringBuf truncatedSuffix)
+{
+    TruncateStringInplace(&string, lengthLimit, truncatedSuffix);
+    return string;
+}
+
+////////////////////////////////////////////////////////////////////////////////
+
 } // namespace NYT
