+++ yt/yt/client/tablet_client/table_mount_cache_detail.cpp
@@ -256,8 +256,8 @@ auto TTableMountCacheBase::TryHandleServantNotActiveError(const TError& error)
         return {};
     }
 
-    if (auto siblingCellDescriptor = attributes.FindYson("sibling_servant_cell_descriptor")) {
-        RegisterCell(ConvertToNode(siblingCellDescriptor));
+    if (auto siblingCellDescriptor = attributes.Find<INodePtr>("sibling_servant_cell_descriptor")) {
+        RegisterCell(std::move(siblingCellDescriptor));
     } else {
         return {};
     }
