+++ yt/yt/core/actions/unittests/future_ut.cpp
@@ -1474,7 +1474,7 @@ TEST_F(TFutureTest, WithDeadlineFail)
     auto deadline = TInstant::Now() + SleepQuantum;
     auto f2 = f1.WithDeadline(deadline);
     EXPECT_EQ(NYT::EErrorCode::Timeout, f2.Get().GetCode());
-    EXPECT_EQ(NYson::ConvertToYsonString(deadline), f2.Get().Attributes().FindYson("deadline"));
+    EXPECT_EQ(deadline, f2.Get().Attributes().Get<TInstant>("deadline"));
 }
 
 TEST_F(TFutureTest, WithTimeoutSuccess)
@@ -1503,7 +1503,7 @@ TEST_F(TFutureTest, WithTimeoutFail)
     auto f1 = p.ToFuture();
     auto f2 = f1.WithTimeout(SleepQuantum);
     EXPECT_EQ(NYT::EErrorCode::Timeout, f2.Get().GetCode());
-    EXPECT_EQ(NYson::ConvertToYsonString(SleepQuantum), f2.Get().Attributes().FindYson("timeout"));
+    EXPECT_EQ(SleepQuantum, f2.Get().Attributes().Get<TDuration>("timeout"));
 }
 
 TEST_W(TFutureTest, Holder)
