+++ yt/yt/core/misc/public.h
@@ -1,7 +1,8 @@
 #pragma once
 
 #include "common.h"
-#include "error_code.h"
+
+#include <library/cpp/yt/error/public.h>
 
 // Google Protobuf forward declarations.
 namespace google::protobuf {
@@ -162,15 +163,6 @@ DECLARE_REFCOUNTED_CLASS(TAdaptiveHedgingManagerConfig)
 DECLARE_REFCOUNTED_STRUCT(IHedgingManager)
 
 ////////////////////////////////////////////////////////////////////////////////
-
-YT_DEFINE_ERROR_ENUM(
-    ((OK)                    (0))
-    ((Generic)               (1))
-    ((Canceled)              (2))
-    ((Timeout)               (3))
-    ((FutureCombinerFailure) (4))
-    ((FutureCombinerShortcut)(5))
-);
 
 DEFINE_ENUM(EProcessErrorCode,
     ((NonZeroExitCode)    (10000))
