+++ yt/yt/core/ytree/convert-inl.h
@@ -182,14 +182,6 @@ T ConstructYTreeConvertibleObject()
 
 ////////////////////////////////////////////////////////////////////////////////
 
-} // namespace NYT::NYTree
-
-////////////////////////////////////////////////////////////////////////////////
-
-namespace NYT::NConvertToImpl {
-
-////////////////////////////////////////////////////////////////////////////////
-
 namespace {
 
 double ConvertYsonStringBaseToDouble(const NYson::TYsonStringBuf& yson)
@@ -234,15 +226,9 @@ TString ConvertYsonStringBaseToString(const NYson::TYsonStringBuf& yson)
 
 ////////////////////////////////////////////////////////////////////////////////
 
-// NB(arkady-e1ppa): TTagInvokeTag uses decltype under the hood
-// meaning the resulting expression is not viable for template argument deduction
-// thus we have to write the type by hand in order to have TTo deducible
-// automatically.
 template <class TTo>
-TTo TagInvoke(NConvertToImpl::TFn<TTo>, const NYTree::INodePtr& node)
+TTo ConvertTo(const NYTree::INodePtr& node)
 {
-    using namespace NYTree;
-
     auto result = ConstructYTreeConvertibleObject<TTo>();
     Deserialize(result, node);
     return result;
@@ -251,10 +237,8 @@ TTo TagInvoke(NConvertToImpl::TFn<TTo>, const NYTree::INodePtr& node)
 ////////////////////////////////////////////////////////////////////////////////
 
 template <class TTo, class TFrom>
-TTo TagInvoke(NConvertToImpl::TFn<TTo>, const TFrom& value)
+TTo ConvertTo(const TFrom& value)
 {
-    using namespace NYTree;
-
     auto type = GetYsonType(value);
     if constexpr (
         NYson::ArePullParserDeserializable<TTo>() &&
@@ -285,9 +269,8 @@ TTo TagInvoke(NConvertToImpl::TFn<TTo>, const TFrom& value)
 
 #define IMPLEMENT_CHECKED_INTEGRAL_CONVERT_TO(type) \
     template <> \
-    inline type TagInvoke(TTagInvokeTag<ConvertTo<type>>, const NYson::TYsonString& str) \
+    inline type ConvertTo(const NYson::TYsonString& str) \
     { \
-        using namespace NYTree; \
         NYson::TTokenizer tokenizer(str.AsStringBuf()); \
         const auto& token = SkipAttributes(&tokenizer); \
         switch (token.GetType()) { \
@@ -316,29 +299,50 @@ IMPLEMENT_CHECKED_INTEGRAL_CONVERT_TO(ui8)
 ////////////////////////////////////////////////////////////////////////////////
 
 template <>
-inline double TagInvoke(TTagInvokeTag<ConvertTo<double>>, const NYson::TYsonString& str)
+inline double ConvertTo(const NYson::TYsonString& str)
 {
     return ConvertYsonStringBaseToDouble(str);
 }
 
 template <>
-inline double TagInvoke(TTagInvokeTag<ConvertTo<double>>, const NYson::TYsonStringBuf& str)
+inline double ConvertTo(const NYson::TYsonStringBuf& str)
 {
     return ConvertYsonStringBaseToDouble(str);
 }
 
 template <>
-inline TString TagInvoke(TTagInvokeTag<ConvertTo<TString>>, const NYson::TYsonString& str)
+inline TString ConvertTo(const NYson::TYsonString& str)
 {
     return ConvertYsonStringBaseToString(str);
 }
 
 template <>
-inline TString TagInvoke(TTagInvokeTag<ConvertTo<TString>>, const NYson::TYsonStringBuf& str)
+inline TString ConvertTo(const NYson::TYsonStringBuf& str)
 {
     return ConvertYsonStringBaseToString(str);
 }
 
 ////////////////////////////////////////////////////////////////////////////////
 
-} // namespace NYT::NConvertToImpl
+} // namespace NYT::NYTree
+
+namespace NYT::NAttributeValueConversionImpl {
+
+////////////////////////////////////////////////////////////////////////////////
+
+template <class T>
+    requires (!CPrimitiveConvertible<T>)
+T TagInvoke(TFrom<T>, const NYson::TYsonString& value)
+{
+    return NYTree::ConvertTo<T>(value);
+}
+
+////////////////////////////////////////////////////////////////////////////////
+
+} // namespace NYT::NAttributeValueConversionImpl
+
+namespace NYT {
+
+using NYT::NYTree::ConvertTo;
+
+} // namespace NYT
