+++ yt/yt/core/ytree/convert.h
@@ -5,8 +5,6 @@
 
 #include <yt/yt/core/yson/consumer.h>
 
-#include <library/cpp/yt/error/convert_to_cpo.h>
-
 namespace NYT::NYson {
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -55,7 +53,11 @@ T ConstructYTreeConvertibleObject();
 
 ////////////////////////////////////////////////////////////////////////////////
 
-using NYT::ConvertTo;
+template <class TTo>
+TTo ConvertTo(const INodePtr& node);
+
+template <class TTo, class TFrom>
+TTo ConvertTo(const TFrom& value);
 
 ////////////////////////////////////////////////////////////////////////////////
