+++ yt/yt/core/ytree/node-inl.h
@@ -4,9 +4,16 @@
 #include "node.h"
 #endif
 
-#include <library/cpp/yt/error/convert_to_cpo.h>
-
 namespace NYT::NYTree {
+
+////////////////////////////////////////////////////////////////////////////////
+
+// Forward declarations
+template <class TTo>
+TTo ConvertTo(const INodePtr& node);
+
+template <class TTo, class TFrom>
+TTo ConvertTo(const TFrom& value);
 
 ////////////////////////////////////////////////////////////////////////////////
